+++ src/Observers/ModelObserver.php
@@ -13,7 +13,7 @@ class ModelObserver
     const DELETED_ACTION = 'Deleted';
     const DELETING_ACTION = 'Deleting';
     const FORCE_DELETED_ACTION = 'Force Deleted';
-    const REPLICATED_ACTION = 'Replicated';
+    const REPLICATING_ACTION = 'Replicating';
     const RESTORED_ACTION = 'Restored';
     const RESTORING_ACTION = 'Restoring';
     const SAVED_ACTION = 'Saved';
@@ -30,66 +30,6 @@ class ModelObserver
             'details' => $details,
             'user_id' => auth()->id()
         ]);
-    }
-    
-    /**
-     * Adds a meta for the given model with the given key and value.
-     * 
-     * @param Model $model
-     * @param string $key
-     * @param string $value
-     */
-    protected function addMeta(Model $model, string $key, string $value)
-    {
-        $model->ledgerMeta()->create([
-            LedgerMeta::META_KEY => $key,
-            LedgerMeta::META_VALUE => $value
-        ]);
-    }
-    
-    /**
-     * Returns the first found LedgerMeta record for the given model with the given key.
-     * 
-     * @param Model $model
-     * @param string $key
-     * @return Model|null
-     */
-    protected function getMeta(Model $model, string $key)
-    {
-        return $model->ledgerMeta()->where(LedgerMeta::META_KEY, $key)->first();
-    }
-
-    /**
-     * Returns the value of the LedgerMeta record found for the given model with the given key.
-     *
-     * @param Model $model
-     * @param string $key
-     * @return mixed|null
-     */
-    protected function getMetaValue(Model $model, string $key)
-    {
-        $meta = $this->getMeta($model, $key);
-
-        if ($meta) {
-            return $meta->meta_value;
-        }
-        
-        return null;
-    }
-    
-    /**
-     * Deletes the first LedgerMeta record found for the given model with the given key.
-     * 
-     * @param Model $model
-     * @param string $key
-     */
-    protected function removeMeta(Model $model, string $key)
-    {
-        $meta = $this->getMeta($model, $key);
-
-        if ($meta) {
-            $meta->delete();
-        }
     }
     
     /**